Extracts ManifestFinderRegistry from ManifestService
