+++ Corgibytes.Freshli.Lib/ManifestService.cs
@@ -9,34 +9,32 @@ namespace Corgibytes.Freshli.Lib
 {
     public class ManifestService
     {
+        // TODO: inject this dependency
         private static readonly Logger logger = LogManager.GetCurrentClassLogger();
         private readonly string _projectRootPath;
-
-        private static readonly IList<AbstractManifestFinder> _finders =
-          new List<AbstractManifestFinder>();
-
-        public static IList<AbstractManifestFinder> Finders => _finders;
 
         public AbstractManifestFinder Finder { get; }
 
         public string[] ManifestFiles =>
-          Finder.GetManifestFilenames(_projectRootPath);
+            Finder.GetManifestFilenames(_projectRootPath);
 
         public bool Successful { get; }
 
         public LibYearCalculator Calculator => new LibYearCalculator(
-          Finder.RepositoryFor(_projectRootPath),
-          Finder.ManifestFor(_projectRootPath)
+            Finder.RepositoryFor(_projectRootPath),
+            Finder.ManifestFor(_projectRootPath)
         );
 
+        // TODO: rework this logic so that it does not run as part of the constructor
         public ManifestService(
-          string projectRootPath,
-          IFileHistoryFinder fileFinder
+            string projectRootPath,
+            IFileHistoryFinder fileFinder
         )
         {
             _projectRootPath = projectRootPath;
             Successful = false;
-            foreach (var finder in Finders.ToImmutableList())
+            // TODO: inject the dependency on ManfestFinderRegistry
+            foreach (var finder in ManifestFinderRegistry.Finders.ToImmutableList())
             {
                 finder.FileFinder = fileFinder;
                 if (finder.GetManifestFilenames(projectRootPath).Any())
@@ -48,47 +46,5 @@ namespace Corgibytes.Freshli.Lib
             }
         }
 
-        public static void Register(AbstractManifestFinder finder)
-        {
-            Finders.Add(finder);
-        }
-
-        public static void RegisterAll()
-        {
-            var manifestFinderTypes = new HashSet<Type>();
-            foreach (var assembly in AppDomain.CurrentDomain.GetAssemblies())
-            {
-                foreach (var type in FindersLoadedIn(assembly))
-                {
-                    manifestFinderTypes.Add(type);
-                }
-            }
-
-            foreach (var type in manifestFinderTypes)
-            {
-                logger.Log(
-                  LogLevel.Info,
-                  $"Registering AbstractManifestFinder: {type}"
-                );
-                Register((AbstractManifestFinder)Activator.CreateInstance(type));
-            }
-        }
-
-        private static IEnumerable<Type> FindersLoadedIn(Assembly assembly)
-        {
-            try
-            {
-                return assembly.GetTypes().
-                  Where(
-                    type => type.BaseType == typeof(AbstractManifestFinder) &&
-                      type.GetConstructor(Type.EmptyTypes) != null
-                  );
-            }
-            catch
-            {
-                logger.Log(LogLevel.Info, $"Unable to load types from {assembly}");
-                return new List<Type>();
-            }
-        }
     }
 }
